Don't re-use existing session identifiers
This fixes a bug where if a user attempted to log in after their session
had already expired, we would resurrect their old session identifier
from the dead.

i.e., if the user logs in once, we would have created a new session and
set the expiry to 30 days in the future

in 30 days, when the session key expires, it will be deleted; however,
the user record will still have a reference to this (now-deleted)
session.

If the user attempted to authenticate again, the old code would have
re-used this session identifier.

To be safe, we now simply create a new session identifier whenever we
successfully authenticate a user.