refactor unit tests for easier login/logout
In particular, there was a race condition wherein the sessions
unit-tests would be creating/deleting a session for the same user as the
other unit tests were attempting to use.

To fix this, we have two 'test users' now in test/conf/test-users.json
which have been pre-authenticated with facebook (manually).  The
ServiceClient class provides a mechanism for logging in with either
user.

Also refactored the 403-authorized routes handling from individual tests
in every single file to a single looping test (that automatically pulls
the authentication-required routes from routes.js).