The PUT /users/:id interface works for usernames
For now, usernames are the only thing we allow the API to modify (per
the service technical design). I can't really remember the use-case for
this API, but I think it had something to do with allowing the user to
change their username from the default (which is their facebook name).