Clean up iOS build after assets cleanup.
This change stabilizes the iOS client build and includes a new
README for how to set up the development environment for iOS.

In particular, the HTML, Javascript, and CSS files have been
updated to reference the relative path '../shared/images' for
images.

The xcodeproj file relies on the 'assets' folder being placed
in the 'client' folder (as a peer of 'html' and 'ios') in order
to find the right image files to use for building.

The README also addresses how to build and initialize the
Facebook iOS SDK when cloning the repository for the first time.